+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Added
 * Add tailwind-based `Tooltip` component.
+* Forward refs to `Button`, `LinkButton` and `CloseButton` tailwind-based components.
 
 ### Changed
 * *Nothing*